data changes and on genetic algorithm

+++ include/sample.h
@@ -6,14 +6,14 @@
 using std::vector;
 
 struct sample{
-	vector<double> input;
+	vector<int> input;
 	int output; // win/lose
 
 	sample();
 
 	sample(vector<int> participants_team_1, vector<int> participants_team_2, int winner);
 
-	void set_input(int i, double val);
+	void set_input(int i, int val);
 
 	void set_output(int out);
 };
